renamed class to a far more appropriate name
